refactor: move url loader from browser to common
